Update README.md
Minor adjustments to text. Removed one sentence since it seemed out of place and is not a recurring element in other documents of this type (i.e. other links in this repo don't detail who the information is for). Changed image path and the link to each subdirectory (previously linked README.md, now links the directory itself).