Document dependency on MIME::Base32
Pull request zonemaster/zonemaster-ldns#178 adds a test-time dependency
on MIME::Base32. This commit documents the change.

This package should be installed manually in build environments.

The module does not seem to be packaged in Red Hat-based Linux
distributions. On those OSes, the MIME::Base32 has to be installed
through CPAN.