Avoid referring to “the zone’s SPF policy”
Technically, any domain, not just zones, can have SPF policies.
Zonemaster will only fetch the domain under test’s SPF policy. This
means that language referring to “the zone’s SPF policy” is technically
incorrect.

This commit updates the following messages tags to remove such language:

 * Z11_INCONSISTENT_SPF_POLICIES;
 * Z11_NO_SPF_FOUND;
 * Z11_SPF1_SYNTAX_ERROR;
 * Z11_SPF1_SYNTAX_OK.

The three last tags in the list get a new argument, {domain}, which is
necessary for the msgid to stay meaningful and user-friendly. The
alternative would be to say something like “the SPF policy at the zone’s
apex”, but “the SPF policy at domain.example” is much better. This is,
however, a change that should ideally be addressed by a database
migration script so that old tests stay meaningful.