Create test zones for Address03
Address03 is missing test zone data. Let’s create it so that we can have
proper unit tests for this test case. It will also help reproduce the
issue described in #1353.

Note that the specification could use a rewrite. I’m basing it on what
we currently have, but a specification update should also include an
update of the test data. Although I think I’m already pretty complete.