Fix undelegated data in Consistency06 test scenario
In the specification for test scenarios for Consistency06, scenario
MULT-SOA-MNAMES-NO-DEL-UNDEL-2 erroneously lists
ns3.mult-soa-mnames-no-del-undel-2.consistency06.xb twice in the
undelegated data. The second occurrence should be ns4, not ns3.

Right above, the description of what SOA MNAME field values appear
mention name servers named ns1 and ns2, where ns3 and ns4 was probably
meant.

Unfortunately for us, fixing the first discrepancy does not fix the
underlying problem: it instead underscores that there is a bug in the
current implementation of Consistency06.