Zone11: handle root, TLDs and .arpa differently
The root zone, TLDs and zones in .arpa are not expected to be domains
appropriate for receiving e-mail. They should not be appropriate as
sender e-mail addresses either.

In the current wording of the specification, a NOTICE (Z11_NO_SPF_FOUND)
is output telling the user that no SPF record was found, when running
Zonemaster on such zones. That message tends to stick out like a sore
thumb, especially on well-configured TLD zones. That NOTICE should not
be there.

The specification is amended so that it outputs specific message tags
when the test is being run on such a zone: if there is no SPF record or
it is a null SPF policy, then we output a special message tag of level
INFO, and if it is a non-null SPF policy, it’s another message of a
higher level.

In doing so, this has the effect of squelching the
NOTICE (Z11_NO_SPF_FOUND) that users have been complaining about when
testing root, TLDs and zones inside .arpa.