MethodsV2: add Get-Parent-NS-Names-and-IPs method
Introduce a new method that obtains the set of name server names and IP
addresses of the parent zone.

Some DNSSEC test cases need to look at information in the parent zone.
There might be problems there, and we need to report the name servers
that the information was obtained from. We’d like to report the name
servers as name server/IP pairs, for example
“ns1.parent.example/198.54.113.1”, like other test cases usually do.

However, it turns out that there is no method that returns the parent’s
name server name/IP pairs. Fortunately, there already exists one that
returns only the IP addresses, named Get-Parent-NS-IP. The outline of
that latter method is reused to define the new method that this commit
introduces, Get-Parent-NS-Names-and-IPs, then Get-Parent-NS-IP is
rewritten in order to take advantage of Get-Parent-NS-Names-and-IPs.

This approach is identical to the way Get-Del-NS-IPs and Get-Zone-NS-IPs
are specified: they use the result of a corresponding
Get-{Del,Zone}-NS-Names-and-IPs method and extract a set of IP addresses
from it.

It does add one more layer of method interdependency, but the extra cost
is negligible: the only penalty is that Get-Parent-NS-IP now iterates
over the result of Get-Parent-NS-Names-and-IPs instead of computing
its result directly.