MethodsV2: update test scenario specification
After introducing Get-Parent-NS-Names-and-IPs, update the test scenario
specification in order to test the method appropriately.

Because Get-Parent-NS-IPs is now defined in terms of
Get-Parent-NS-Names-and-IPs, only the latter needs to be tested. But the
existing scenarios only tested the former.

This commit modifies the test zone specification, and updates the
expected output for MethodsV2 to list parent NS names and IP addresses
instead of only parent NS IP addresses.