Fix outline in MethodsV2 test scenario spec
There was a little mixup in the section heading levels in the test
scenarios: the first scenarios used level 2 heading, then moved on to
level 3 headings.

This is fixed by using the test scenario specification Address03 as a
template and moving all scenarios into level 3 headings.

The table of contents at the beginning also needed an update.

As a bonus change, clean up whitespace in the .md file, using Emacs’s
M-x whitespace-cleanup command.