Fix delegation inconsistency in test arpa. zone
The parent (root zone) delegates arpa. to ns1. and ns2., but the NS
records at the apex of the arpa. zone listed different name servers:
ns1.xa and ns2.xa. This inconsistency caused some odd errors when trying
to run zonemaster-cli on those test zones, while preparing the
test-zones-output.md.

This might be the sign of a bug somewhere in Basic02, which should
warrant further investigation.